chore: drop /v1 subpath export from env, functions, ai-sdk-provider
The package root already re-exported the full `/v1` surface, so the
versioned subpath was redundant for these libraries. Consolidate each
`src/v1.ts` into `src/index.ts`, remove the `./v1` export + build entry,
and point docs/tests at the package root.

Versioned subpath exports remain on @neondatabase/config and
@neondatabase/config-runtime, where pinning a policy-schema major is
meaningful.

Minor (0.x) bumps via changeset — pre-GA, so not treated as a major:
env 0.7.0 → 0.8.0, functions 0.2.0 → 0.3.0, ai-sdk-provider 0.4.0 → 0.5.0.
