+++ packages/ai-sdk-provider/e2e/gateway-matrix.e2e.test.ts
@@ -7,7 +7,7 @@ import {
 } from "ai";
 import { describe, expect, it } from "vitest";
 import { z } from "zod";
-import { neon } from "../src/v1.js";
+import { neon } from "../src/index.js";
 import {
 	assertGatewayEnv,
 	expectNoHardFailureWarnings,
