+++ packages/ai-sdk-provider/package.json
@@ -1,6 +1,6 @@
 {
 	"name": "@neondatabase/ai-sdk-provider",
-	"version": "0.4.0",
+	"version": "0.5.0",
 	"description": "Community Vercel AI SDK provider for the Neon AI Gateway.",
 	"keywords": [
 		"neon",
@@ -29,11 +29,6 @@
 			"types": "./dist/index.d.ts",
 			"import": "./dist/index.js",
 			"default": "./dist/index.js"
-		},
-		"./v1": {
-			"types": "./dist/v1.d.ts",
-			"import": "./dist/v1.js",
-			"default": "./dist/v1.js"
 		}
 	},
 	"files": [
